+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added documentation for the new `terra-file-path` component.
+
 ## 1.39.0 - (October 3, 2023)
 
 * Added